+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+- executeAsyncOperation method
+
 ## [0.1.0] - 2021-11-03
 
 ### Added